// generated/orval/fetch/google-health-api/models/stageSummaryType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Output only. Sleep stage type: AWAKE, DEEP, REM, LIGHT etc.
 */
export type StageSummaryType = typeof StageSummaryType[keyof typeof StageSummaryType];


export const StageSummaryType = {
  SLEEP_STAGE_TYPE_UNSPECIFIED: 'SLEEP_STAGE_TYPE_UNSPECIFIED',
  AWAKE: 'AWAKE',
  LIGHT: 'LIGHT',
  DEEP: 'DEEP',
  REM: 'REM',
  ASLEEP: 'ASLEEP',
  RESTLESS: 'RESTLESS',
} as const;